backend/requirements.txt
Flask==2.3.2
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.1
ollama==0.5.1
openai==1.84.0
fastapi[standard]==0.112.0